Rename java package
