+++ src/main/kotlin/io/github/linesegments/api/SimpleLineSegmentsService.kt
@@ -1,9 +1,9 @@
-package io.github.simonedavico.api
+package io.github.linesegments.api
 
-import io.github.simonedavico.segments.Line
-import io.github.simonedavico.segments.Point
-import io.github.simonedavico.segments.Space
-import io.github.simonedavico.segments.findLineSegments
+import io.github.linesegments.segments.Line
+import io.github.linesegments.segments.Point
+import io.github.linesegments.segments.Space
+import io.github.linesegments.segments.findLineSegments
 import io.vertx.core.Future
 
 /**
