--- src/main/kotlin/io/github/simonedavico/RestServer.kt
@@ -1,8 +1,0 @@
-package io.github.simonedavico
-
-import io.github.simonedavico.api.ApiUtils
-import io.github.simonedavico.api.RestApi
-
-fun main(args: Array<String>) {
-    ApiUtils.runVerticle(RestApi::class.java)
-}
