+++ src/main/resources/output.md
@@ -1,0 +1,84 @@
+# 5 points
+
+`[]`
+
+# 4 points
+
+```
+[ [ {
+  "x" : -1,
+  "y" : 0
+}, {
+  "x" : 1,
+  "y" : 2
+}, {
+  "x" : 3,
+  "y" : 4
+}, {
+  "x" : 4,
+  "y" : 5
+} ] ]
+```
+
+# 3 points
+
+```
+[ [ {
+  "x" : 5,
+  "y" : 0
+}, {
+  "x" : 3,
+  "y" : 4
+}, {
+  "x" : 4,
+  "y" : 2
+} ], [ {
+  "x" : 0,
+  "y" : 6
+}, {
+  "x" : 2,
+  "y" : 4
+}, {
+  "x" : 4,
+  "y" : 2
+} ], [ {
+  "x" : 4,
+  "y" : 6
+}, {
+  "x" : 4,
+  "y" : 5
+}, {
+  "x" : 4,
+  "y" : 2
+} ], [ {
+  "x" : 3,
+  "y" : 4
+}, {
+  "x" : -1,
+  "y" : 0
+}, {
+  "x" : 4,
+  "y" : 5
+}, {
+  "x" : 1,
+  "y" : 2
+} ], [ {
+  "x" : 4,
+  "y" : 6
+}, {
+  "x" : 0,
+  "y" : -2
+}, {
+  "x" : 3,
+  "y" : 4
+} ], [ {
+  "x" : 3,
+  "y" : 7
+}, {
+  "x" : 0,
+  "y" : -2
+}, {
+  "x" : 2,
+  "y" : 4
+} ] ]
+```
