+++ src/test/kotlin/io/github/linesegments/api/SimpleLineSegmentsServiceTest.kt
@@ -1,10 +1,10 @@
-package io.github.simonedavico.api
+package io.github.linesegments.api
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import com.fasterxml.jackson.module.kotlin.readValue
-import io.github.simonedavico.segments.Point
+import io.github.linesegments.segments.Point
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
